[FEAT] Member 및 Course Entity 및 service 제작중 #3

+++ src/main/java/com/lms/domain/Course/dto/CourseDTO.java
@@ -11,7 +11,7 @@ public class CourseDTO{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long no;     // 기본키
+    private Integer no;     // 기본키
 
     @Enumerated(EnumType.STRING)
     private Subject subject;        //과정
